ai_secretary/requirements.txt
openai
Flask
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
